Split search logic into a new file to avoid a too big single file controller
